backend/src/all_ai_interview-main/database_generate/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.9.2
python-multipart==0.0.6
openai
python-dotenv==1.0.0
